pyproject.toml
[project]
name = "basicmcp"
version = "0.1.6"
description = "Model context protocol plugins"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipython>=8.32.0",
    "lancedb>=0.19.0",
    "mcp[cli]==1.3.0",
    "ruff>=0.9.7",
    "sentence-transformers>=3.4.1",
    "flask>=3.0.0",
    "openai>=1.12.0",
    "tree-sitter==0.21.3",
    "tree_sitter_languages",
    "cohere>=4.48",
    "pandas>=2.2.0",
    "redis>=5.0.1",
    "markdown>=3.5.2",
    "tiktoken>=0.6.0",
    "rerankers",
    "rerankers[transformers]",
    "python-dotenv>=1.0.1",
    "open-clip-torch",
    "pillow",
    "gitpython>=3.1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0",
]

[[project.authors]]
name = "ayush"
email = "ayush.chaurarsia@gmail.com"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
basicmcp-codeqa = "basicmcp:main"
basicmcp = "basicmcp:dummy"

[tool.pytest.ini_options]
addopts = "--cov=basicmcp --cov-report=xml --cov-report=term-missing"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src/basicmcp"]

